added BaseLoginActivity class
